Refactor code to add spawner
